feat(desktop): schema-certified env scaffolding for agent snapshots
Unify env/credential handling across snapshots, runs, and chat-completions
with a schema-certified contract path:

- definition.environment exports env key NAMES as import scaffolding
  (sorted, reserved/malformed filtered); values stay non-serializable.
- definition.environmentValues accepts import-only value hints from
  external producers (e.g. a control plane shipping an OpenAI-compatible
  API route). Secret-named keys are dropped to blank scaffolding,
  undeclared keys ignored, count/per-value capped at decode; Buzz export
  never writes the field. Preview surfaces pre-filled keys.
- Run-scoped env overlay merges into agent_runtime_env for declared keys
  across every run mode, incl. contract-less/legacy imports.
- agent_runtime_env is the single consumer-visible truth surface for
  declared env values; contract terms fire only when schemas exist.
- env.get_credential resolves through the central credential-name
  normalizer (fail-closed on ambiguous input).
- Persona and loop export/import propagate scaffolding; the env vars
  module remains the only keeper of secret-bearing literals.
- egress_guard inventories follow the persona-import module split.

cargo test --lib: 2021 passed, 0 failed
cargo clippy --lib --all-features: clean
cargo fmt --check: clean
