+++ desktop/src-tauri/src/commands/personas/snapshot.rs
@@ -479,6 +479,8 @@ mod png_body_tests {
                 name_pool: vec![],
                 idle_timeout_seconds: None,
                 max_turn_duration_seconds: None,
+                environment: vec![],
+                environment_values: Default::default(),
             },
             profile: crate::managed_agents::agent_snapshot::AgentSnapshotProfile {
                 display_name: "Agent".to_string(),
