+++ desktop/src-tauri/src/commands/personas/snapshot/fidelity_tests.rs
@@ -88,6 +88,8 @@ fn make_snapshot(
             name_pool: vec![],
             idle_timeout_seconds: None,
             max_turn_duration_seconds: None,
+            environment: vec![],
+            environment_values: Default::default(),
         },
         profile: AgentSnapshotProfile {
             display_name: "Test Agent".to_string(),
