+++ desktop/src-tauri/src/commands/personas/snapshot/tests.rs
@@ -108,6 +108,8 @@ fn make_snapshot(
             name_pool: vec![],
             idle_timeout_seconds: None,
             max_turn_duration_seconds: None,
+            environment: vec![],
+            environment_values: Default::default(),
         },
         profile: AgentSnapshotProfile {
             display_name: "Test Agent".to_string(),
