+++ desktop/src/testing/e2eBridge.ts
@@ -10881,6 +10881,7 @@ export function maybeInstallE2eTauriMocks() {
           memoryEntryCount: 0,
           hasSourceAllowlist: false,
           sourceAllowlistCount: 0,
+          environmentKeys: [],
         };
       }
       case "confirm_agent_snapshot_import": {
